almost done game state

+++ item.ml
@@ -1,5 +1,5 @@
 
-type t = {item_name: string; item_value: int}
+type t = {item_name: string; item_value: float}
 
 let make_item name value = 
   {item_name = name; item_value = value}
